Move Navigation and Root components to containers components directory.
