+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
-import Root from './components/Root';
+import Root from './containers/Root';
 import configureStore from './redux/store';
 import './reset.css';
 import './index.css';
